requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
herpetologist==0.0.9
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.9.0
llvmlite==0.38.0
malaya-boilerplate==0.0.15
malaya-speech==1.2.3
Markdown==3.3.6
matplotlib==3.5.1
memoization==0.4.0
noisereduce==2.0.0
numba==0.55.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-speech-features==0.6
python-vlc==3.0.12118
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.62.3
typing-extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
